+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.1] - 2022-10-27
+### Added
+- GLC Parser
+  - `pile::Parser::Grammar` parser<file> function
+  - unit test for `pile::Parser::Grammar` parser<file> function
+
+### Changed
+- Assets now have a test subdirectory for unit tests files
 
 ## [0.1.0] - 2022-10-07
 ### Added